fix invalid test names generation
